requirements.txt
Django==4.2.3
requests
psycopg2>=2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.2.0
django-filter==23.2
pika
django-environ==0.9.0
pytest-cov==4.1.0
pytest-django==4.5.2
mixer==7.2.2
channels
channels-redis